add tests for step_manager_service
shift order of id/missing term/error/warning handling in process job
so that id is always output first (and you know any term, error,
warning handling messages are for the id above it when watching
processing)

call finalize_processing_report with file, not report object, to
decouple a bit for testing. It turns out I could not figure out how to
test this method here, but I'm leaving the change.