use handler.service_type instead of digging into config
`cspace-batch-import` should not have to know about the internals of
the mapper structure. `collectionspace-mapper` is already all up in
the mapper structure and should be able to tell the importer what it
needs to know cleanly through the DataHandler interface.

`RecordTransferService` still has a couple of mapper excavations to
get rid of, but they are not immediate problems, so they'll be on the
list for later improvements