Attach report files to step as job running
If a Sidekiq process fails currently all record of the job activity
is lost. By attaching the files incrementally while the job is
running we can provide the user with the reports near to the point
of failure (near because we are not attaching for every row
processed but along with the checkin point for status updates).

The checkin point may need to become configurable.