Improve error message for invalid character encoding in CSV
- As long as CSV is getting invalid_encoding error(s), only the new
message will be shown. This is because encoding errors sometimes cause
structural errors that cascade through the file, causing hundreds of
error messages about structure in which the message about encoding
gets lost. Often if the encoding is fixed, the structural errors
disappear. This makes it obvious/clear that fixing the invalid
character encoding is the required.
- Shows the first occurrence of each invalid character in context, so
users can more easily find/fix the issues

Improve error message for invalid character encoding in CSV

- As long as CSV is getting invalid_encoding error(s), only the new
message will be shown. This is because encoding errors sometimes cause
structural errors that cascade through the file, causing hundreds of
error messages about structure in which the message about encoding
gets lost. Often if the encoding is fixed, the structural errors
disappear. This makes it obvious/clear that fixing the invalid
character encoding is the required.
- Shows the first occurrence of each invalid character in context, so
users can more easily find/fix the issues