Remove within_csv_row_limit? method
The method is in ApplicationController, but is only used in
BatchesController. That makes it hard to find.

If we create a new csv_length_ok? method in BatchesController (which I
wanted to do anyway because CSV length is not part of Csvlint validity
being ok), then it is only used once.

A method for simple logic only used once seems like overkill.